+++ src/main/java/redstonedubstep/mods/vanishmod/misc/SoundSuppressionHelper.java
@@ -2,8 +2,8 @@ package redstonedubstep.mods.vanishmod.misc;
 
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 
 import org.apache.commons.lang3.tuple.Pair;
 
@@ -12,10 +12,12 @@ import net.minecraft.network.protocol.Packet;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.projectile.Projectile;
 import net.minecraft.world.level.GameType;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.ChestBlock;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.Vec3;
 import net.minecraft.world.phys.shapes.BooleanOp;
@@ -66,74 +68,126 @@ public class SoundSuppressionHelper {
 		return packetOrigin != null && packetOrigin.getLeft().equals(packet) ? packetOrigin.getRight() : null;
 	}
 
-	public static boolean shouldSuppressSoundEventFor(Player player, Level level, double x, double y, double z, Player forPlayer) {
-		return shouldSuppressSoundEventFor(player, level, new Vec3(x, y, z), forPlayer);
+	public static Player getIndirectVanishedSoundCause(Player player, Level level, double x, double y, double z, Player forPlayer) {
+		return getIndirectVanishedSoundCause(player, level, new Vec3(x, y, z), forPlayer);
 	}
 
 	//Returns true if a vanished player directly produced the sound, or if it is determined that a vanished player was indirectly causing a sound, and that it thus should not be broadcast
-	public static boolean shouldSuppressSoundEventFor(Player player, Level level, Vec3 soundOrigin, Player forPlayer) {
+	public static Player getIndirectVanishedSoundCause(Player player, Level level, Vec3 soundOrigin, Player forPlayer) {
 		if (player != null)
-			return VanishUtil.isVanished(player, forPlayer);
+			return VanishUtil.isVanished(player, forPlayer) ? player : null;
 
 		if (!VanishConfig.CONFIG.indirectSoundSuppression.get())
-			return false;
+			return null;
 
-		return SoundSuppressionHelper.areVanishedPlayersAt(level, soundOrigin, forPlayer) || SoundSuppressionHelper.vanishedPlayerVehicleAt(level, soundOrigin, forPlayer) || SoundSuppressionHelper.vanishedPlayersInteractWith(level, BlockPos.containing(soundOrigin), forPlayer);
+		Player vanishedSoundCause = SoundSuppressionHelper.getVanishedPlayerInteractedWith(level, soundOrigin, forPlayer);
+
+		if (vanishedSoundCause == null)
+			vanishedSoundCause = SoundSuppressionHelper.getVanishedPlayerAt(level, soundOrigin, forPlayer);
+
+		if (vanishedSoundCause == null)
+			vanishedSoundCause = SoundSuppressionHelper.getVanishedProjectileOwnerAt(level, soundOrigin, forPlayer);
+
+		if (vanishedSoundCause == null)
+			vanishedSoundCause = SoundSuppressionHelper.getVanishedPlayerWithVehicleAt(level, soundOrigin, forPlayer);
+
+		return vanishedSoundCause;
 	}
 
 	//Returns true if a vanished player directly produced the sound, or if it is determined that a vanished player was indirectly causing a sound, and that it thus should not be broadcast
-	public static boolean shouldSuppressSoundEventFor(Player player, Level level, Entity soundOrigin, Player forPlayer) {
+	public static Player getIndirectVanishedSoundCause(Player player, Level level, Entity soundOrigin, Player forPlayer) {
 		if (player != null)
-			return VanishUtil.isVanished(player, forPlayer);
+			return VanishUtil.isVanished(player, forPlayer) ? player : null;
 
 		if (!VanishConfig.CONFIG.indirectSoundSuppression.get() || soundOrigin == null)
-			return false;
+			return null;
 
-		return SoundSuppressionHelper.areVanishedPlayersAt(level, soundOrigin.position(), forPlayer) || SoundSuppressionHelper.isVanishedPlayerVehicle(soundOrigin, forPlayer) || SoundSuppressionHelper.vanishedPlayersInteractWith(level, soundOrigin, forPlayer);
+		Player vanishedSoundCause = SoundSuppressionHelper.getVanishedPlayerInteractedWith(level, soundOrigin, forPlayer);
+
+		if (vanishedSoundCause == null)
+			vanishedSoundCause = SoundSuppressionHelper.getVanishedPlayerAt(level, soundOrigin.position(), forPlayer);
+
+		if (vanishedSoundCause == null)
+			vanishedSoundCause = SoundSuppressionHelper.getVanishedProjectileOwner(soundOrigin, forPlayer);
+
+		if (vanishedSoundCause == null)
+			vanishedSoundCause = SoundSuppressionHelper.getVanishedPlayerInVehicle(soundOrigin, forPlayer);
+
+		return vanishedSoundCause;
 	}
 
-	public static boolean shouldSuppressParticlesFor(Player player, Level level, double x, double y, double z, Player forPlayer) {
+	public static Player getIndirectVanishedParticleCause(Player player, Level level, double x, double y, double z, Player forPlayer) {
 		Vec3 particleOrigin = new Vec3(x, y, z);
 
 		if (player != null)
-			return VanishUtil.isVanished(player, forPlayer);
+			return VanishUtil.isVanished(player, forPlayer) ? player : null;
 
 		if (!VanishConfig.CONFIG.indirectParticleSuppression.get())
-			return false;
+			return null;
 
-		return SoundSuppressionHelper.areVanishedPlayersAt(level, particleOrigin, forPlayer) || SoundSuppressionHelper.vanishedPlayerVehicleAt(level, particleOrigin, forPlayer) || SoundSuppressionHelper.vanishedPlayersInteractWith(level, BlockPos.containing(particleOrigin), forPlayer);
+		Player vanishedParticleCause = SoundSuppressionHelper.getVanishedPlayerInteractedWith(level, particleOrigin, forPlayer);
+
+		if (vanishedParticleCause == null)
+			vanishedParticleCause = SoundSuppressionHelper.getVanishedPlayerAt(level, particleOrigin, forPlayer);
+
+		if (vanishedParticleCause == null)
+			vanishedParticleCause = SoundSuppressionHelper.getVanishedProjectileOwnerAt(level, particleOrigin, forPlayer);
+
+		if (vanishedParticleCause == null)
+			vanishedParticleCause = SoundSuppressionHelper.getVanishedPlayerWithVehicleAt(level, particleOrigin, forPlayer);
+
+		return vanishedParticleCause;
 	}
 
-	public static boolean areVanishedPlayersAt(Level level, Vec3 pos, Player forPlayer) {
+	public static Player getVanishedPlayerAt(Level level, Vec3 pos, Player forPlayer) {
 		VoxelShape shape = Shapes.block().move(pos.x - 0.5D, pos.y - 0.5D, pos.z - 0.5D);
-		return vanishedPlayersAndHitResults.keySet().stream().anyMatch(p -> p.level().equals(level) && p.gameMode.getGameModeForPlayer() != GameType.SPECTATOR && VanishUtil.isVanished(p, forPlayer) && Shapes.joinIsNotEmpty(shape, Shapes.create(p.getBoundingBox()), BooleanOp.AND));
+
+		return vanishedPlayersAndHitResults.keySet().stream().filter(p -> p.level().equals(level) && p.gameMode.getGameModeForPlayer() != GameType.SPECTATOR && VanishUtil.isVanished(p, forPlayer) && Shapes.joinIsNotEmpty(shape, Shapes.create(p.getBoundingBox()), BooleanOp.AND)).findFirst().orElse(null);
 	}
 
-	public static boolean vanishedPlayerVehicleAt(Level level, Vec3 pos, Player forPlayer) {
+	public static Player getVanishedProjectileOwnerAt(Level level, Vec3 pos, Player forPlayer) {
+		AABB soundArea = AABB.ofSize(pos, 1.0D, 1.0D, 1.0D);
+		List<Entity> projectiles = level.getEntities((Entity) null, soundArea, e -> e instanceof Projectile);
+
+		if (!projectiles.isEmpty())
+			return vanishedPlayersAndHitResults.keySet().stream().filter(p -> p.level().equals(level) && p.gameMode.getGameModeForPlayer() != GameType.SPECTATOR && VanishUtil.isVanished(p, forPlayer) && projectiles.stream().anyMatch(proj -> ((Projectile) proj).getOwner().equals(p))).findFirst().orElse(null);
+
+		return null;
+	}
+
+	public static Player getVanishedProjectileOwner(Entity entity, Player forPlayer) {
+		if (entity instanceof Projectile projectile)
+			return vanishedPlayersAndHitResults.keySet().stream().filter(p -> p.gameMode.getGameModeForPlayer() != GameType.SPECTATOR && VanishUtil.isVanished(p, forPlayer) && projectile.getOwner().equals(p)).findFirst().orElse(null);
+
+		return null;
+	}
+
+	public static Player getVanishedPlayerWithVehicleAt(Level level, Vec3 pos, Player forPlayer) {
 		VoxelShape shape = Shapes.block().move(pos.x - 0.5D, pos.y - 0.5D, pos.z - 0.5D);
-		return vanishedPlayersAndHitResults.keySet().stream().filter(p -> p.level().equals(level) && p.gameMode.getGameModeForPlayer() != GameType.SPECTATOR && VanishUtil.isVanished(p, forPlayer)).map(Entity::getVehicle).filter(Objects::nonNull).anyMatch(v -> Shapes.joinIsNotEmpty(shape, Shapes.create(v.getBoundingBox()), BooleanOp.AND));
+
+		return vanishedPlayersAndHitResults.keySet().stream().filter(p -> p.level().equals(level) && p.gameMode.getGameModeForPlayer() != GameType.SPECTATOR && VanishUtil.isVanished(p, forPlayer)).map(p -> Pair.of(p, p.getVehicle())).filter(pv -> pv.getRight() != null && Shapes.joinIsNotEmpty(shape, Shapes.create(pv.getRight().getBoundingBox()), BooleanOp.AND)).findFirst().map(Pair::getLeft).orElse(null);
 	}
 
-	public static boolean isVanishedPlayerVehicle(Entity entity, Player forPlayer) {
-		return vanishedPlayersAndHitResults.keySet().stream().anyMatch(p -> p.gameMode.getGameModeForPlayer() != GameType.SPECTATOR  && VanishUtil.isVanished(p, forPlayer) && entity.equals(p.getVehicle()));
+	public static Player getVanishedPlayerInVehicle(Entity entity, Player forPlayer) {
+		return vanishedPlayersAndHitResults.keySet().stream().filter(p -> p.gameMode.getGameModeForPlayer() != GameType.SPECTATOR && VanishUtil.isVanished(p, forPlayer) && entity.equals(p.getVehicle())).findFirst().orElse(null);
 	}
 
-	public static boolean vanishedPlayersInteractWith(Level level, BlockPos pos, Player forPlayer) {
-		return vanishedPlayersAndHitResults.entrySet().stream().anyMatch(e -> e.getKey().level().equals(level) && VanishUtil.isVanished(e.getKey(), forPlayer) && e.getValue() != null && equalsThisOrConnected(pos, level, e.getValue().getLeft()));
+	public static Player getVanishedPlayerInteractedWith(Level level, Vec3 pos, Player forPlayer) {
+		return vanishedPlayersAndHitResults.entrySet().stream().filter(e -> e.getKey().level().equals(level) && VanishUtil.isVanished(e.getKey(), forPlayer) && e.getValue() != null && equalsThisOrConnected(pos, level, e.getValue().getLeft())).findFirst().map(Map.Entry::getKey).orElse(null);
 	}
 
-	public static boolean vanishedPlayersInteractWith(Level level, Entity entity, Player forPlayer) {
-		return vanishedPlayersAndHitResults.entrySet().stream().anyMatch(e -> e.getKey().level().equals(level) && VanishUtil.isVanished(e.getKey(), forPlayer) && e.getValue() != null && entity.equals(e.getValue().getRight()));
+	public static Player getVanishedPlayerInteractedWith(Level level, Entity entity, Player forPlayer) {
+		return vanishedPlayersAndHitResults.entrySet().stream().filter(e -> e.getKey().level().equals(level) && VanishUtil.isVanished(e.getKey(), forPlayer) && e.getValue() != null && entity.equals(e.getValue().getRight())).findFirst().map(Map.Entry::getKey).orElse(null);
 	}
 
-	public static boolean equalsThisOrConnected(BlockPos soundPos, Level level, BlockPos interactPos) {
-		if (soundPos.equals(interactPos))
-			return true;
-		else if (interactPos != null) {
+	public static boolean equalsThisOrConnected(Vec3 soundPos, Level level, BlockPos interactPos) {
+		if (interactPos != null) {
 			BlockState state = level.getBlockState(interactPos);
 
-			if (state.getBlock() instanceof ChestBlock)
-				return soundPos.equals(interactPos.relative(ChestBlock.getConnectedDirection(state)));
+			if (new AABB(interactPos).inflate(0.01).contains(soundPos)) //expand the interaction pos a little to cover (literal) edge cases
+				return true;
+			else if (state.getBlock() instanceof ChestBlock)
+				return new AABB(interactPos.relative(ChestBlock.getConnectedDirection(state))).inflate(0.01).contains(soundPos);
 		}
 
 		return false;