+++ Changelog.md
@@ -1,6 +1,9 @@
 -----------Changelog for v1.1.17-----------
 
 - Added the "/v help" subcommand as an ingame explanation of all subcommands in this mod
+- Added tracing, which is a debug functionality that reports any event caused by you that has been concealed for other players
+- Added sound suppressing for projectiles owned by a vanished player
+- Fixed the vanish command feedback message being able to reveal to admins that a player has vanished
 
 -----------Changelog for v1.1.16-----------